std::vector in brakepipe

+++ simulator/brakepipe/include/sweep.h
@@ -13,13 +13,15 @@
  * \date 06/11/2016
  */
 
-#ifndef SWEEP_H
-#define SWEEP_H
+#ifndef     SWEEP_H
+#define     SWEEP_H
+
+#include    <cstddef>
 
 /*!
  * \fn
  * \brief Sweep method of linear equation system solve
  */
-void sweep(int n, double *A, double *B, double *C, double *b, double *x);
+void sweep(size_t n, double *A, double *B, double *C, double *b, double *x);
 
 #endif //SWEEP_H
